fix: changes temperature to integer

+++ RTD_firmware/Core/Inc/ADS1248.h
@@ -20,9 +20,9 @@ void ADS124X_init(SPI_HandleTypeDef *hspi,
         GPIO_TypeDef *GPIO_START, uint16_t PIN_START,
         GPIO_TypeDef *GPIO_CS, uint16_t PIN_CS);
 
-int32_t Temperature(SPI_HandleTypeDef *hspi,
+float Temperature(SPI_HandleTypeDef *hspi,
         GPIO_TypeDef *GPIO_CS, uint16_t PIN_CS);
 
-int32_t RTD_Converter(int32_t resistance);
+float RTD_Converter(int32_t resistance);
 
 #endif /* INC_ADS1248_H_ */
